+++ PoKeysLibWS2812.c
@@ -47,7 +47,7 @@ int32_t PK_WS2812_SendLEDdataEx(sPoKeysDevice* device, uint32_t * LEDdata, uint1
     return PK_OK;
 }
 
-int32_t PK_WS2812_SendLEDdata(sPoKeysDevice* device, uint32_t * LEDdata, uint16_t startLED, uint16_t LEDcount)
+int32_t PK_WS2812_SendLEDdata(sPoKeysDevice* device, uint32_t * LEDdata, uint16_t startLED, uint8_t LEDcount)
 {
     int32_t errCode = PK_OK;
     uint16_t offset = 0;
